### 2021-05-03 #3235 @Kevin-Mattheus-Moerman
@PhilBoileau thanks for this submission. Can you work on checking/fixing those missing/invalid DOI's? :point_up:
If you make changes to the paper you can update it here by calling:
`@whedon generate pdf`
