### 2021-05-03 #3235 @fboehm
@PhilBoileau - Do you have any suggestions for reviewers? These can be people who have reviewed for JOSS in the past or any scholars in your area. If you would leave their names and github usernames in the comment thread, I'll then reach out to them.
