+++ R/plotFunctions.R
@@ -1,3 +1,0 @@
-#' Functions for Plotting cvCovEst Results
-#'
-
