### 2021-05-04 #3235 @fboehm
:wave: @wrathematics & @AnestisTouloumis, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
