+++ paper/paper.bib
@@ -129,7 +129,7 @@
   pages = {365-411},
   year = {2004},
   issn = {0047-259X},
-  doi = {https://doi.org/10.1016/S0047-259X(03)00096-4},
+  doi = {10.1016/S0047-259X(03)00096-4},
   url = {http://www.sciencedirect.com/science/article/pii/S0047259X03000964},
   author = {Ledoit, Olivier and Wolf, Michael}
 }
