### 2021-05-06 0eca17d
cleaned up citations and added missing dois
