+++ paper/paper.bib
@@ -1,28 +1,28 @@
 @article{fan2016,
-    author = {Fan, Jianqing and Liao, Yuan and Liu, Han},
-    title = {An overview of the estimation of large covariance and precision
-      matrices},
-    journal = {The Econometrics Journal},
-    volume = {19},
-    number = {1},
-    pages = {C1-C32},
-    doi = {10.1111/ectj.12061},
-    url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/ectj.12061},
-    year = {2016}
+  author = {Fan, Jianqing and Liao, Yuan and Liu, Han},
+  title = {An overview of the estimation of large covariance and precision
+           matrices},
+  journal = {The Econometrics Journal},
+  volume = {19},
+  number = {1},
+  pages = {C1-C32},
+  doi = {10.1111/ectj.12061},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/ectj.12061},
+  year = {2016}
 }
 
 @techreport{laan_dudoit:2003,
-  author = {Dudoit, Sandrine and {van der Laan}, Mark J},
-  title = {Asymptotics of Cross-Validated Risk Estimation in Estimator
-    Selection and Performance Assessment},
+  author = {{van der Laan}, Mark J and Dudoit, Sandrine },
+  title = {Unified Cross-Validation Methodology For Selection Among Estimators
+           and a General Cross-Validated Adaptive Epsilon-Net Estimator: Finite Sample Oracle Inequalities and Examples},
   series = {U.C. Berkeley Division of Biostatistics Working Paper Series},
   type = {Working Paper},
-  number = {126},
-  month = {February},
+  number = {130},
+  month = {November},
   year = {2003},
   institution = {University of California, Berkeley},
   address = {Berkeley},
-  URL = {https://biostats.bepress.com/ucbbiostat/paper126}
+  URL = {https://biostats.bepress.com/ucbbiostat/paper130}
 }
 
 @article{dudoit2005,
@@ -34,6 +34,7 @@
   number={2},
   pages={131--154},
   year={2005},
+  doi = {10.1016/j.stamet.2005.02.003},
   publisher={Elsevier}
 }
 
@@ -44,7 +45,8 @@
   journal={Statistics and Decisions},
   year={2006},
   volume={24},
-  pages={351-371}
+  pages={351-371},
+  doi = {10.1524/stnd.2006.24.3.351}
 }
 
 @manual{R,
@@ -56,7 +58,7 @@
   url = {https://www.R-project.org/}
 }
 
-@Misc{future,
+@misc{future,
   author = {Bengtsson, Henrik},
   title = {A Unifying Framework for Parallel and Distributed Processing in R
     using Futures},
@@ -148,14 +150,18 @@
   year = {2010}
 }
 
-@techreport{Ledoit2020,
-  title = {Analytical nonlinear shrinkage of large-dimensional covariance
-    matrices},
-  author = {Ledoit, Olivier and Wolf, Michael},
-  year = {2018},
-  institution = {Department of Economics - University of Zurich},
-  number = {264},
-  url = {https://EconPapers.repec.org/RePEc:zur:econwp:264}
+@article{Ledoit2020,
+  author = {Olivier Ledoit and Michael Wolf},
+  title = {{Analytical nonlinear shrinkage of large-dimensional covariance matrices}},
+  volume = {48},
+  journal = {The Annals of Statistics},
+  number = {5},
+  publisher = {Institute of Mathematical Statistics},
+  pages = {3043 -- 3065},
+  keywords = {Hilbert transform, Kernel estimation, rotation equivariance},
+  year = {2020},
+  doi = {10.1214/19-AOS1921},
[…]
