+++ paper/paper.bib
@@ -266,7 +266,7 @@
   keywords = {empirical orthogonal functions, Fredholm determinant, Karhunen–Loève transform, Laguerre ensemble, Laguerre polynomial, Largest eigenvalue, largest singular value, Liouville–Green method, Painlevé equation, Plancherel–Rotach asymptotics, Random matrix theory, Tracy–Widom distribution, Wishart distribution},
   year = {2001},
   doi = {10.1214/aos/1009210544},
-  URL = {https://doi.org/10.1214/aos/1009210544}r={JSTOR}
+  URL = {https://doi.org/10.1214/aos/1009210544}
 }
 
 @article{tasic2016,
