### 2021-05-06 593153b
adding citation for UMAP example
