### 2021-05-11 #3273 @whedon
[REVIEW]: cvCovEst: Cross-validated covariance matrix estimator selection and evaluation in R
**Submitting author:** @PhilBoileau (<a href="http://orcid.org/0000-0002-4850-2507">Philippe Boileau</a>)
**Repository:** <a href="https://github.com/PhilBoileau/cvCovEst" target ="_blank">https://github.com/PhilBoileau/cvCovEst</a>
**Version:** 1.0.0
