### 2021-06-15 #3273 @Marie-PerrotDockes
Hello,

I'm so sorry I will start the review right now!!

Just to be sure : I am reviewing the code but no the paper ? (I just check
the paper to get the idea of the methods but that is all ? )

Sorry for the delay,
Marie

Le dim. 6 juin 2021 à 21:12, Frederick Boehm ***@***.***> a
écrit :

> @yunanwu123 <https://github.com/yunanwu123> and @Marie-PerrotDockes
> <https://github.com/Marie-PerrotDockes> - Since JOSS reviews are a little
> different from those of many journals, please feel free to ask questions if
> you're unsure about anything. Thanks again!
>
