### 2021-06-15 PhilBoileau/cvCovEst#49 @Marie-PerrotDockes
Add tests on the values return by the function
Hello,

Their are already many tests available to check your package (nice ;)) but I feel most of them are checking the arguments, and that the functions run without leaving messages or errors but can you add test to check if the value return by the function are the one that are expected ? For instance, for a given seed can you manually check that you find the same empirical_risk than in your function for the different estimators ?

Best,
