### 2021-06-19 PhilBoileau/cvCovEst#49 @PhilBoileau
Indeed, we have accumulated many tests -- sorry if it was overwhelming! Thank you for your suggestion. While we already have a handful of estimator-specific and loss function tests to verify that they are calculated correctly, we have added a few more tests for our loss functions to verify their accuracy in `tests/testthat/test-foldFunctions.R`.Unfortunately, the empirical `emprirical_risk` column in the `risk_df` tibble output by `cvCovEst()` corresponds to the cross-validated risk estimate. We've updated the name of this variable to `cv_risk` to avoid confusion. However, since this is the cross-validated risk, it is much more difficult to manually verify. Further, this cross-validated risk estimate is computed by the `origami` package under the hood. We trust that they've tested their package to ensure that these estimates are accurate.
