### 2021-06-19 f325572
bumped the version number and recorded changes in NEWS.md
