### 2021-06-19 #3273 @PhilBoileau
Here are the links to the issues, @fboehm: PhilBoileau/cvCovEst#49, PhilBoileau/cvCovEst#49. @Marie-PerrotDockes, if you're satisfied with my responses and updates, feel free to close the issues!
