### 2021-06-23 PhilBoileau/cvCovEst#53 @yunanwu123
Report summary tables & Add non-Gaussian examples
- The plot.cvCovEst function reports the eigenvalues plot, risk plot, heatmap plot and the summary table with the best performing estimator within each class. The first three plots can be exported separately. Is it possible to output the summary table (in latex format, or other formats) separately? I believe it would be useful in presenting the analysis results. You could use "xtable" or other packages.
- It would be more convincing to add a non-Gaussian example in the reports, such as Gamma distributions.
- A typo: the argument in plot.cvCovEst for data is "dat_orig". However, in the **_Toy Dataset Example_** in your paper, it is "data_in".
