### 2021-07-04 2feb0f7
removing groups from summary tables
