+++ NAMESPACE
@@ -102,7 +102,6 @@ importFrom(rlang,quo_get_expr)
 importFrom(stats,median)
 importFrom(stats,optimize)
 importFrom(stats,quantile)
-importFrom(stats,t.test)
 importFrom(stats,var)
 importFrom(stringr,str_split)
 importFrom(stringr,str_squish)
