### 2021-07-04 abf57a3
summary() now return a table instead of a list when a single summary function is provided
