+++ R/summary_functions.R
@@ -247,7 +247,10 @@ summary.cvCovEst <- function(
     return(f)
   })
 
-  names(out) <- sums_to_exec
+  if (length(out) == 1)
+    out <- out[[1]]
+  else
+    names(out) <- sums_to_exec
 
   return(out)
 }
