### 2021-07-04 0d350af
bumping version number and updating NEWS
