### 2021-07-04 PhilBoileau/cvCovEst#53 @PhilBoileau
Thank you for these suggestions! I'll address them in order:

1. That's a good idea. I've updated `summary.cvCovEst()` to return a `tibble` when a single summarization function is specified by the `summ_fun` argument. Users can now pass this `tibble` to the presentation function of their choice, (e.g. `xtable::xtable()`, `knitr::kable()`, etc). I've avoided implementing a specific table formatting function so that users have more flexibility.
2. Employing Gaussian data, which is familiar to most, seems reasonable for illustration purposes. We also provide a real-data example. If readers remain unconvinced, they can review our [preprint](https://arxiv.org/abs/2102.09715) on the method. A much more comprehensive simulation study is performed in this work.
3. Thank you for your sharp eye. It's been fixed.

Thank you for reviewing our software! Please let me know if there is anything else that you would like me address.
