### 2021-07-10 #3273 @fboehm
Here are some needed revisions to the references:

1. Levina's name is misspelled in two refs. It should be "Levina", not "Levita".
2. R should be capitalized in the Bengttson ref.
3. You use both "The Annals of Statistics" and "Annals of Statistics" in your references. Did the journal change its name? Or is this an error? If there was no name change, then please update the references to reflect the true journal name.
4. The book from Pourahmadi presumably has a doi. Please display it in the reference. Or correct me if I'm mistaken.
5. The R package nlshrink should be in all lower case letters, I believe. It looks like the title of this ref has the "n" capitalized.
