### 2021-07-10 #3273 @fboehm
Here are some needed revisions for the text of the manusript:

1. Figure 1 legend: "estimators" is missing an apostrophe; should be "estimator's", I think.
2. lines 150 - 151: Should "Recherche" and "Technologies" be capitalized? I'm not sure.
