### 2021-07-10 #3273 @fboehm
One minor addition - please cite the `pkgdown` R package in your manuscript, where you mention it near the end of the article.
