### 2021-07-10 af30b6d
added pkgdown citation and addressed spelling mistakes
