+++ paper/paper.bib
@@ -314,3 +314,11 @@ Covariance Matrices},
     note = {R package version 1.0.1},
     url = {https://CRAN.R-project.org/package=nlshrink},
   }
+
+@Manual{pkgdown,
+  title = {pkgdown: Make Static HTML Documentation for a Package},
+  author = {Hadley Wickham and Jay Hesselberth},
+  year = {2020},
+  note = {R package version 1.6.1},
+  url = {https://CRAN.R-project.org/package=pkgdown}
+}
