+++ paper/paper.bib
@@ -60,7 +60,7 @@
 
 @misc{future,
   author = {Bengtsson, Henrik},
-  title = {A Unifying Framework for Parallel and Distributed Processing in R
+  title = {A Unifying Framework for Parallel and Distributed Processing in {R}
     using Futures},
   year = {2020},
   month = {aug},
@@ -95,7 +95,7 @@
 }
 
 @article{Bickel2008_thresh,
-  author = {Bickel, Peter J and Levita, Elizaveta},
+  author = {Bickel, Peter J and Levina, Elizaveta},
   doi = {10.1214/08-AOS600},
   journal = {Annals of Statistics},
   month = {12},
@@ -109,7 +109,7 @@
 }
 
 @article{bickel2008_banding,
-  author = {Bickel, Peter J and Levita, Elizaveta},
+  author = {Bickel, Peter J and Levina, Elizaveta},
   doi = {10.1214/009053607000000758},
   journal = {Annals of Statistics},
   month = {02},
@@ -154,7 +154,7 @@
   author = {Olivier Ledoit and Michael Wolf},
   title = {{Analytical nonlinear shrinkage of large-dimensional covariance matrices}},
   volume = {48},
-  journal = {The Annals of Statistics},
+  journal = {Annals of Statistics},
   number = {5},
   publisher = {Institute of Mathematical Statistics},
   pages = {3043 -- 3065},
@@ -253,13 +253,14 @@
   Series = {Wiley series in probability and statistics},
   Title = {High-dimensional covariance estimation},
   Year = {2013},
+  ISBN = {978-1-118-03429-3}
 }
 
 @article{johnstone2001distribution,
   author = {Iain M. Johnstone},
   title = {{On the distribution of the largest eigenvalue in principalcomponents analysis}},
   volume = {29},
-  journal = {The Annals of Statistics},
+  journal = {Annals of Statistics},
   number = {2},
   publisher = {Institute of Mathematical Statistics},
   pages = {295 -- 327},
@@ -307,7 +308,7 @@
   }
 
 @manual{ramprasad2016,
-    title = {nlshrink: Non-Linear Shrinkage Estimation of Population Eigenvalues and
+    title = {{nlshrink}: Non-Linear Shrinkage Estimation of Population Eigenvalues and
 Covariance Matrices},
     author = {Pratik Ramprasad},
     year = {2016},
