### 2021-07-10 #3273 @PhilBoileau
Alright, the following issues have been addressed:
- "Levita" has been changed to "Levina"
- "R" is capitalized in the Bengttson reference
- "The Annals of Statistics" has been changed to "Annals of Statistics"
- The ISBN for Pourehmadi's book has been included in the bibtex reference
- `nlshrink`'s title is now fixed in the bibliography
- The typo in Figure 1's caption has been fixed
- `pkgdown` is now cited where appropriate

Note that the capitalization of "Fonds de recherche du Québec - Nature et technologies" is the correct scheme in french.

Thank you again for identifying these errors, and for editing this review! I'll archive the package as soon as you give me the signal.
