### 2021-07-12 #3273 @fboehm
Thanks for making the changes, @PhilBoileau ! One tiny additional fix is to the term "pkgdown" in the references. Can you fix the bib file so that the "p" is lower case?

Thanks again!
