+++ paper/paper.bib
@@ -292,7 +292,7 @@
 }
 
 @manual{lee2019,
-    title = {CovTools: Statistical Tools for Covariance Analysis},
+    title = {{CovTools}: Statistical Tools for Covariance Analysis},
     author = {Kyoungjae Lee and Kisung You},
     year = {2019},
     note = {R package version 0.5.3},
@@ -300,7 +300,7 @@
   }
 
 @manual{wang2014,
-    title = {CVTuningCov: Regularized Estimators of Covariance Matrices with CV Tuning},
+    title = {{CVTuningCov}: Regularized Estimators of Covariance Matrices with CV Tuning},
     author = {Binhuan Wang},
     year = {2014},
     note = {R package version 1.0},
@@ -316,8 +316,8 @@ Covariance Matrices},
     url = {https://CRAN.R-project.org/package=nlshrink},
   }
 
-@Manual{pkgdown,
-  title = {pkgdown: Make Static HTML Documentation for a Package},
+@manual{pkgdown,
+  title = {{pkgdown}: Make Static HTML Documentation for a Package},
   author = {Hadley Wickham and Jay Hesselberth},
   year = {2020},
   note = {R package version 1.6.1},
