### 2021-07-24 cee00f5
update to docs before preprint pkg version release
