+++ inst/REFERENCES.bib
@@ -58,7 +58,7 @@
 }
 
 @article{Bickel2008_thresh,
-  author = {Peter J. Bickel and Elizaveta Levita},
+  author = {Peter J. Bickel and Elizaveta Levina},
   doi = "10.1214/08-AOS600",
   journal = "Annals of Statistics",
   month = "12",
@@ -72,7 +72,7 @@
 }
 
 @article{bickel2008_banding,
-  author = {Peter J. Bickel and Elizaveta Levita},
+  author = {Peter J. Bickel and Elizaveta Levina},
   doi = "10.1214/009053607000000758",
   journal = "Annals of Statistics",
   month = "02",
@@ -219,7 +219,7 @@
 }
 
 @misc{boileau2021,
-      title={Cross-Validated Loss-Based Covariance Matrix Estimator Selection in High Dimensions}, 
+      title={Cross-Validated Loss-Based Covariance Matrix Estimator Selection in High Dimensions},
       author={Philippe Boileau and Nima S. Hejazi and Mark J. van der Laan and Sandrine Dudoit},
       year={2021},
       eprint={2102.09715},
