### 2021-07-24 2d5ef9e
removed ability to pass in true covariance matrix, updated docs
