+++ R/foldFunctions.R
@@ -18,12 +18,6 @@
 #'  hyperparameter(s). These hyperparameters may be in the form of a single
 #'  \code{numeric} or a \code{numeric} vector. If no hyperparameter is needed
 #'  for a given estimator, then the estimator need not be listed.
-#' @param true_cov_mat A \code{matrix} like object representing the true
-#'  covariance matrix of the data generating distribution. This parameter
-#'  is intended for use only in simulation studies, and defaults to a value
-#'  of \code{NULL}. If not null, \code{cvFrobeniusLoss} returns the true
-#'  cross-validated Frobenius loss in addition to the cross-validated Frobenius
-#'  loss estimate for the given \code{fold}.
 #'
 #' @importFrom dplyr bind_rows slice
 #' @importFrom origami training validation fold_index
@@ -69,9 +63,12 @@
 #'   )
 #' )
 #' @export
-cvFrobeniusLoss <- function(fold, dat, estimator_funs,
-                            estimator_params = NULL,
-                            true_cov_mat = NULL) {
+cvFrobeniusLoss <- function(
+  fold,
+  dat,
+  estimator_funs,
+  estimator_params = NULL
+) {
 
   # split the data into training and validation
   train_data <- origami::training(dat)
@@ -121,32 +118,13 @@ cvFrobeniusLoss <- function(fold, dat, estimator_funs,
       est_square <- matrixStats::sum2(est_mat^2)
 
       # return the results from the fold
-      if (is.null(true_cov_mat)) {
-        out <- tibble::tibble(
-          estimator = est_name,
-          hyperparameters = estimator_hparam,
-          loss = 1 / nrow(valid_data) * (elwise_sq - 2 * had_crossprod) +
-            est_square,
-          fold = origami::fold_index(fold = fold)
-        )
-      } else {
-
-        # fit the covariance matrix estimator on the full dataset
-        # NOTE: this is here out of convenience...it is not computationally
-        # efficient, but not all that important since only for simulations,
-        # and this will not be released in the main package.
-        est_mat_full <- est_fun(dat)
-
-        out <- tibble::tibble(
-          estimator = est_name,
-          hyperparameters = estimator_hparam,
-          loss = 1 / nrow(valid_data) * (elwise_sq - 2 * had_crossprod) +
-            est_square,
-          true_loss = trueFrobeniusLoss(est_mat, true_cov_mat),
-          full_mat_loss = trueFrobeniusLoss(est_mat_full, true_cov_mat),
-          fold = origami::fold_index(fold = fold)
-        )
-      }
+      out <- tibble::tibble(
+        estimator = est_name,
+        hyperparameters = estimator_hparam,
+        loss = 1 / nrow(valid_data) * (elwise_sq - 2 * had_crossprod) +
+          est_square,
+        fold = origami::fold_index(fold = fold)
+      )
 
       return(out)
     } else {
@@ -179,35 +157,14 @@ cvFrobeniusLoss <- function(fold, dat, estimator_funs,
           est_square <- matrixStats::sum2(est_mat^2)
 
           # return the results from the fold
-          if (is.null(true_cov_mat)) {
-            out <- tibble::tibble(
-              estimator = est_name,
-              hyperparameters = estimator_hparam,
-              loss = 1 / nrow(valid_data) * (elwise_sq - 2 * had_crossprod) +
-                est_square,
-              fold = origami::fold_index(fold = fold)
-            )
-          } else {
-            # fit the covariance matrix estimator on the full dataset
-            # NOTE: this is here out of convenience...it is not computationally
-            # efficient, but not all that important since only for simulations,
[…]
