+++ R/cvCovEst.R
@@ -37,13 +37,6 @@
 #' @param parallel A \code{logical} option indicating whether to run the main
 #'  cross-validation loop with \code{\link[future.apply]{future_lapply}()}. This
 #'  is passed directly to \code{\link[origami]{cross_validate}()}.
-#' @param true_cov_mat A \code{matrix}-like object giving the true covariance
-#'  matrix of the data-generating distribution, which is assumed Gaussian. This
-#'  parameter is intended as an aid for use only in simulation studies, and it
-#'  defaults to \code{NULL}. When not \code{NULL}, various conditional risk
-#'  difference ratios of the estimator selected by \code{\link{cvCovEst}()} are
-#'  computed relative to the different oracle selectors. NOTE: This parameter
-#'  will be phased out by the release of version 1.0.0.
 #'
 #' @importFrom origami cross_validate folds_vfold folds_montecarlo
 #' @importFrom dplyr arrange summarise group_by "%>%" ungroup
@@ -66,20 +59,6 @@
 #'     \item \code{args} - A named \code{list} containing arguments passed to
 #'       \code{cvCovEst}.
 #'   }
-#'   If the true covariance matrix is input through \code{true_cov_mat}, then
-#'   three additional elements are returned:
-#'   \itemize{
-#'     \item \code{cv_cv_riskdiff} - A \code{numeric} corresponding to the
-#'       conditional cross-validated risk difference of the cvCovEst selection.
-#'     \item \code{oracle_cv_riskdiff} - A \code{numeric} corresponding to the
-#'       conditional cross-validated risk difference of the oracle selection.
-#'     \item \code{cv_oracle_riskdiff_ratio} - A \code{numeric} corresponding
-#'       to the cross-validated risk difference ratio of the cvCovEst selection
-#'       and the cross-validated dataset oracle selection.
-#'     \item \code{full_data_riskdiff_ratio} - A \code{numeric} corresponding
-#'       to the full data risk difference ratio of the cvCovEst selection and
-#'       the full dataset oracle selection.
-#'   }
 #'
 #' @examples
 #' cvCovEst(
@@ -95,20 +74,20 @@
 #' )
 #' @export
 cvCovEst <- function(
-                     dat,
-                     estimators = c(
-                       linearShrinkEst, thresholdingEst, sampleCovEst
-                     ),
-                     estimator_params = list(
-                       linearShrinkEst = list(alpha = 0),
-                       thresholdingEst = list(gamma = 0)
-                     ),
-                     cv_loss = cvMatrixFrobeniusLoss,
-                     cv_scheme = "v_fold", mc_split = 0.5, v_folds = 10L,
-                     center = TRUE,
-                     scale = FALSE,
-                     parallel = FALSE,
-                     true_cov_mat = NULL) {
+  dat,
+  estimators = c(
+   linearShrinkEst, thresholdingEst, sampleCovEst
+  ),
+  estimator_params = list(
+   linearShrinkEst = list(alpha = 0),
+   thresholdingEst = list(gamma = 0)
+  ),
+  cv_loss = cvMatrixFrobeniusLoss,
+  cv_scheme = "v_fold", mc_split = 0.5, v_folds = 10L,
+  center = TRUE,
+  scale = FALSE,
+  parallel = FALSE
+) {
 
   # grab estimator expression
   estimators <- rlang::enquo(estimators)
@@ -133,8 +112,7 @@ cvCovEst <- function(
[…]
