### 2021-07-24 3f136be
updating cvCovEst version stated in paper
