### 2021-07-24 #3273 @PhilBoileau
Apologies for the delay, @fboehm! I've made a new release and archived it on Zenodo.  The package version is 1.0.0, and its DOI is [10.5281/zenodo.5132903](https://zenodo.org/record/5132903#.YPxeDZNKg-Q).
