### 2023-10-06 #3273 @ajstewartlang
:wave: @PhilBoileau Might you have the bandwidth to help out with reviewing the following submission please? I suspect it overlaps quite well with your expertise and interests:

https://github.com/openjournals/joss-reviews/issues/5862
