### 2023-10-08 #3273 @PhilBoileau
I would have loved to, but I’m friends with the first author. I suspect that goes against JOSS’s review policy. I’d be happy to review other articles for JOSS, however.
